Separate code by header file and other code review changes
