+++ src/lib.rs
@@ -19,14 +19,18 @@ pub use keys::*;
 pub use serial::*;
 pub use types::*;
 pub use usb::*;
-pub use hid_manager::*;
+pub use io_hid_keys::*;
+pub use io_hid_base::*;
+pub use io_hid_manager::*;
 
 mod io_return;
 mod keys;
 mod serial;
 mod types;
 mod usb;
-mod hid_manager;
+mod io_hid_keys;
+mod io_hid_base;
+mod io_hid_manager;
 
 // exports from <IOKit/IOKitLib.h>
